Split out host message metrics conversion
To keep source files focused on the same type of conversion.

Updated the signature of `PopulateCuttlefishHost` to take a
`MetricsEventV2` reference instead of a `CuttlefishHost` reference to
have a clean boundary where the split out conversion files handle the
submessage entirely.

Bug: 549999548
