+++ base/cvd/cuttlefish/host/libs/metrics/BUILD.bazel
@@ -139,6 +139,23 @@ cf_cc_library(
 )
 
 cf_cc_library(
+    name = "host_conversion",
+    srcs = ["host_conversion.cc"],
+    hdrs = ["host_conversion.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:host_info",
+        "//cuttlefish/host/libs/metrics:gce_environment",
+        "//cuttlefish/host/libs/metrics:github_environment",
+        "//cuttlefish/host/libs/metrics:host_metrics",
+        "//cuttlefish/host/libs/metrics:invoker",
+        "//external_proto:cf_gce_environment_cc_proto",
+        "//external_proto:cf_github_actions_environment_cc_proto",
+        "//external_proto:cf_host_cc_proto",
+        "//external_proto:cf_metrics_event_v2_cc_proto",
+    ],
+)
+
+cf_cc_library(
     name = "host_metrics",
     srcs = ["host_metrics.cc"],
     hdrs = ["host_metrics.h"],
@@ -163,22 +180,16 @@ cf_cc_library(
     srcs = ["metrics_conversion.cc"],
     hdrs = ["metrics_conversion.h"],
     deps = [
-        "//cuttlefish/common/libs/utils:host_info",
         "//cuttlefish/host/libs/config:data_image_policy",
         "//cuttlefish/host/libs/config:gpu_mode",
         "//cuttlefish/host/libs/metrics:device_event_type",
         "//cuttlefish/host/libs/metrics:fetch_conversion",
         "//cuttlefish/host/libs/metrics:fetch_metrics",
-        "//cuttlefish/host/libs/metrics:gce_environment",
-        "//cuttlefish/host/libs/metrics:github_environment",
         "//cuttlefish/host/libs/metrics:guest_metrics",
+        "//cuttlefish/host/libs/metrics:host_conversion",
         "//cuttlefish/host/libs/metrics:host_metrics",
-        "//cuttlefish/host/libs/metrics:invoker",
         "//external_proto:cf_flags_cc_proto",
-        "//external_proto:cf_gce_environment_cc_proto",
-        "//external_proto:cf_github_actions_environment_cc_proto",
         "//external_proto:cf_guest_cc_proto",
-        "//external_proto:cf_host_cc_proto",
         "//external_proto:cf_log_cc_proto",
         "//external_proto:cf_metrics_event_v2_cc_proto",
         "@fmt",
